Completed Board and GameSquare
Done Board and GameSquare. All function in document added. Attributes all added except for the width and length replaced it with just a size since its a square. No tests yet but added a function that prints all of the board coordinates and the piece occupying the gamesquare. Enumeration of board theme will depend on who color is used on the UI.