Testing and fixing possible moves
Added temp if statement in game so it ignores errors
Add get and set for both row and col in gameSquare.py for convinience sake
Added build_chess_board in board.py for creation of board
Changed rook for possibleMoves.py
Some piece move now need to move from the current piece selected to each corner. to stop at the right time if a piece is blocking a path.