Fixed the way build_list_of_moves was being called in Player.py and removed the commented out code.
In build list of moves for all pieces - Brian re wrote it this because the way it was written before did not properly trigger boolean short circuiting so it blows up on the get_colour() for a None type

In PossibleMoves.py, I fixed some spacing so it follows PEP 8.  And deleted some unnecessary comment delineating methods not in the domain model.  Comments like these just go out of date fast.

I took out the method to return the list of build moves, this is not needed now that we are not object oriented here.

I also moved the select best method into it's own file.  I think we will need this eventually to chose the best one from the list of possible destination squares.